requirements.txt
asgiref==3.10.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.8
django-allauth==0.57.2
django-summernote==0.8.20.0
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.1
webencodings==0.5.1
whitenoise==6.11.0
